requirements.txt
django==3.2.6
djangorestframework==3.12.4
django-rest-swagger==2.2.0
psycopg2==2.9.1
pyyaml==5.4.1
